Extract SignersManager class
